chore: extract creating job to a function
